# web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/guilds/{guildId}/ai-settings": {
      "get": {
        "operationId": "AiSettingsController_read",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildAiSettingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      },
      "put": {
        "operationId": "AiSettingsController_update",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuildAiSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildAiSettingsDto"
                }
              }
            }
          },
          "403": {
            "description": "Managing this server is required."
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/credentials/{provider}": {
      "put": {
        "operationId": "AiSettingsController_putCredential",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutCredentialDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Key stored, encrypted at rest."
          },
          "503": {
            "description": "The secret vault is not configured on this instance."
          }
        },
        "tags": [
          "AiSettings"
        ]
      },
      "delete": {
        "operationId": "AiSettingsController_removeCredential",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key removed."
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/credentials/{provider}/test": {
      "post": {
        "operationId": "AiSettingsController_testCredential",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialTestResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/transcription": {
      "get": {
        "operationId": "AiSettingsController_readTranscription",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionSettingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      },
      "put": {
        "operationId": "AiSettingsController_updateTranscription",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionSettingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/transcription/auth-token": {
      "put": {
        "operationId": "AiSettingsController_putTranscriptionAuthToken",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutWakeSecretDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Token stored, encrypted at rest."
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/transcription/models": {
      "get": {
        "operationId": "AiSettingsController_remoteTranscriptionModels",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteWhisperModelsDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/transcription/test": {
      "post": {
        "operationId": "AiSettingsController_testTranscription",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionProbeDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/transcription/wake": {
      "post": {
        "operationId": "AiSettingsController_wakeTranscription",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionProbeDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
...
    "/api/v1/guilds/{guildId}/ai-settings/wake-secrets/{method}/{field}": {
      "put": {
        "operationId": "AiSettingsController_putWakeSecret",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutWakeSecretDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Stored, encrypted at rest."
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/session-estimate": {
      "get": {
        "operationId": "AiSettingsController_sessionEstimate",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minutes",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCostEstimateDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/pricing": {
      "get": {
        "operationId": "AiSettingsController_pricingOverrides",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PricingOverrideDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      },
      "put": {
        "operationId": "AiSettingsController_savePricingOverrides",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePricingOverridesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PricingOverrideDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
    "/api/v1/guilds/{guildId}/ai-settings/models": {
      "get": {
        "operationId": "AiSettingsController_models",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModelsDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
...
    "/api/v1/campaigns/{campaignId}/ai-settings/embedding/estimate": {
      "get": {
        "operationId": "AiSettingsController_estimateReindex",
        "parameters": [
          {
            "name": "model",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexEstimateDto"
                }
              }
            }
          }
        },
        "tags": [
          "AiSettings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lestapenna API",
    "description": "Read-only campaign data for the authenticated web app.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lp_session"
      }
    },
    "schemas": {
...
      "UpdateGuildAiSettingsDto": {
        "type": "object",
        "properties": {
          "quality": {
            "nullable": true,
            "description": "Null clears the choice and falls back to the instance defaults.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TierChoiceDto"
              }
            ]
          },
          "fast": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TierChoiceDto"
              }
            ]
          },
          "image": {
            "nullable": true,
            "description": "The model that draws portraits. Null clears the choice.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TierChoiceDto"
              }
            ]
          }
        }
      },
      "PutCredentialDto": {
        "type": "object",
        "properties": {
          "api_key": {
            "type": "string",
            "description": "The API key. Write-only: no endpoint ever returns it."
          }
        },
        "required": [
          "api_key"
        ]
      },
...
    }
  }
}